txm 0428

+++ src/com/entity/Goods.java
@@ -9,7 +9,7 @@ public class Goods {
     private String goodsId;
     private String goodsName;
     private String goodsDetails;
-    private Integer goodsPrice;
+    private Double goodsPrice;
     private String goodsStoreid;
     private Integer goodsKucun;
     private Integer goodsSell;
@@ -22,7 +22,7 @@ public class Goods {
     }
 
     //商品详情的构造方法,有"商家名"
-    public Goods(String goodsStoreid,String storeName,String goodsId,String goodsName,String goodsDetails, Integer goodsPrice, Integer goodsKucun, Integer goodsSell, String goodsStyle, String goodsPicture) {
+    public Goods(String goodsStoreid,String storeName,String goodsId,String goodsName,String goodsDetails, Double goodsPrice, Integer goodsKucun, Integer goodsSell, String goodsStyle, String goodsPicture) {
         this.goodsStoreid = goodsStoreid;
         this.goodsId = goodsId;
         this.storeName = storeName;
